:sparkles: feat: add feature flag setup
