+++ .env.example
@@ -4,7 +4,8 @@ NODE_ENV=development
 NEXT_PUBLIC_VERCEL_ENV=development
 NEXT_PUBLIC_VERCEL_URL=localhost:3000
 
-SENTRY_AUTH_TOKEN=
+NEXT_PUBLIC_SENTRY_TUNNEL=/monitoring
+SENTRY_DSN=
 
 #SUPABASE
 NEXT_PUBLIC_SUPABASE_URL=
@@ -25,4 +26,6 @@ KEEPALIVE_API_KEY=
 
 #ANALYTICS
 NEXT_PUBLIC_POSTHOG_KEY=
-NEXT_PUBLIC_POSTHOG_HOST=
+NEXT_PUBLIC_POSTHOG_HOST=https://app.posthog.com
+NEXT_PUBLIC_POSTHOG_ENABLE_LOCAL=false
+POSTHOG_API_KEY=
