+++ src/lib/services/tracking/feature-flag-keys.ts
@@ -1,0 +1,3 @@
+export const FEATURE_FLAG_KEYS = {
+  betaCodeAnimate: 'beta-code-animate',
+} as const
