+++ src/lib/services/tracking/index.ts
@@ -3,6 +3,7 @@
 import posthog from 'posthog-js'
 
 import { ensurePostHogClient } from '../../posthog'
+import { FEATURE_FLAG_KEYS } from './feature-flag-keys'
 
 export const analytics = {
     identify: (userId: string, traits?: Record<string, any>) => {
@@ -21,3 +22,5 @@ export const analytics = {
         posthog.capture(eventName, properties)
     },
 }
+
+export { FEATURE_FLAG_KEYS }
